requirements.txt
Django==6.0
python-dotenv==1.0.1
django-allauth==65.4.0
requests==2.32.3
google-generativeai==0.8.4
Pillow==12.0.0
PyJWT==2.10.1
cryptography==46.0.3
django-extensions==3.2.3
gunicorn==23.0.0
resend==2.19.0
